todo-7-module: refactoring. divide blocks for modules, services and components. Removed parseStorageValues from app.js
