+++ scripts/components/Task/generate-template.component.js
