+++ scripts/components/Task/renderer.js
@@ -1,0 +1,3 @@
+export function renderer(root, block) {
+    root.append(block);
+}
