--- scripts/modules/renderer.module.js
@@ -1,3 +1,0 @@
-export function rendererModule(root, block) {
-    root.append(block);
-}
